Move test to statistics-test
