+++ src/parquet/CMakeLists.txt
@@ -52,7 +52,6 @@ ADD_PARQUET_TEST(column_writer-test)
 ADD_PARQUET_TEST(properties-test)
 ADD_PARQUET_TEST(statistics-test)
 ADD_PARQUET_TEST(encoding-test)
-ADD_PARQUET_TEST(parquet_reader_writer-test)
 ADD_PARQUET_TEST(public-api-test)
 ADD_PARQUET_TEST(types-test)
 ADD_PARQUET_TEST(reader-test)
